api-cursos-django-main/requirements.txt
2captcha-python==1.2.0
asgiref==3.8.1
attrs==25.3.0
bcrypt==3.2.0
beautifulsoup4==4.13.3
blinker==1.9.0
capsolver==1.0.6
certifi==2025.1.31
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.4
cryptography==3.4.8
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
fasteners==0.14.1
Flask==3.1.0
flask-cors==5.0.1
future==0.18.2
h11==0.14.0
httplib2==0.20.2
icoextract==0.1.2
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==3.0.2
monotonic==1.6
more-itertools==8.10.0
nextcaptcha-python @ git+https://github.com/nextcaptcha/nextcaptcha-python.git@6a1bf72845f5e36c137a3b9e4d1cfd859e56769b
numpy==2.2.4
oauthlib==3.2.0
olefile==0.46
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
paramiko==2.9.3
pefile==2021.9.3
pexpect==4.8.0
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pyelftools==0.27
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-debian==0.1.43
python-dotenv==1.1.0
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.1
requests==2.32.3
SecretStorage==3.3.1
selenium==4.11.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.5
wadllib==1.3.6
webdriver-manager==4.0.0
Werkzeug==3.1.3
wsproto==1.2.0
xdg==5
zipp==1.0.0